+++ web/src/popovers.js
@@ -6,7 +6,7 @@ import * as emoji_picker from "./emoji_picker";
 import * as overlays from "./overlays";
 import * as playground_links_popover from "./playground_links_popover";
 import * as popover_menus from "./popover_menus";
-import * as resize from "./resize";
+import * as right_sidebar_ui from "./right_sidebar_ui";
 import * as stream_popover from "./stream_popover";
 import * as user_card_popover from "./user_card_popover";
 import * as user_group_popover from "./user_group_popover";
@@ -110,15 +110,6 @@ export function focus_first_action_popover_item() {
     focus_first_popover_item($items);
 }
 
-export function hide_userlist_sidebar() {
-    $(".app-main .column-right").removeClass("expanded");
-}
-
-export function show_userlist_sidebar() {
-    $(".app-main .column-right").addClass("expanded");
-    resize.resize_page_components();
-}
-
 // On mobile web, opening the keyboard can trigger a resize event
 // (which in turn can trigger a scroll event).  This will have the
 // side effect of closing popovers, which we don't want.  So we
@@ -198,7 +189,7 @@ export function hide_all_except_sidebars(opts) {
 // This function will hide all the popovers, including the mobile web
 // or narrow window sidebars.
 export function hide_all(not_hide_tippy_instances) {
-    hide_userlist_sidebar();
+    right_sidebar_ui.hide_userlist_sidebar();
     stream_popover.hide_streamlist_sidebar();
     hide_all_except_sidebars({
         not_hide_tippy_instances,
