+++ web/src/user_search.js
@@ -2,6 +2,7 @@ import $ from "jquery";
 
 import * as popovers from "./popovers";
 import * as resize from "./resize";
+import * as right_sidebar_ui from "./right_sidebar_ui";
 import * as stream_popover from "./stream_popover";
 
 export class UserSearch {
@@ -99,7 +100,7 @@ export class UserSearch {
             if ($column.hasClass("column-left")) {
                 stream_popover.show_streamlist_sidebar();
             } else if ($column.hasClass("column-right")) {
-                popovers.show_userlist_sidebar();
+                right_sidebar_ui.show_userlist_sidebar();
             }
         }
     }
